+++ core/genesis.go
@@ -323,6 +323,7 @@ func (g *Genesis) toBlock(db ethdb.Database, triedb *triedb.Database) *types.Blo
 		// Granite: set TimeMilliseconds
 		if confExtra.IsGranite(g.Timestamp) {
 			headerExtra.TimeMilliseconds = new(uint64)
+			*headerExtra.TimeMilliseconds = g.Timestamp * 1000
 		}
 	}
