+++ plugin/evm/customheader/extra.go
@@ -13,7 +13,6 @@ import (
 	"github.com/ava-labs/libevm/core/types"
 
 	"github.com/ava-labs/coreth/params/extras"
-	"github.com/ava-labs/coreth/plugin/evm/customtypes"
 	"github.com/ava-labs/coreth/plugin/evm/upgrade/acp176"
 	"github.com/ava-labs/coreth/plugin/evm/upgrade/ap0"
 	"github.com/ava-labs/coreth/plugin/evm/upgrade/ap3"
@@ -172,49 +171,6 @@ func VerifyExtra(rules extras.AvalancheRules, extra []byte) error {
 			)
 		}
 	}
-	return nil
-}
-
-// VerifyTime verifies that the header's Time and TimeMilliseconds fields are
-// consistent with the given rules and the current time.
-// This includes:
-// - Time is not too far in the future
-// - TimeMilliseconds is nil before Granite activation
-// - TimeMilliseconds is non-nil after Granite activation
-// - Time matches TimeMilliseconds/1000 after Granite activation
-// - (TODO) Time is strictly increasing if parent is also after Granite activation
-// - (TODO) Minimum block delay is enforced
-// - (TODO) More time-related checks from dummy.VerifyHeader
-func VerifyTime(extraConfig *extras.ChainConfig, header *types.Header, currentTime time.Time) error {
-	headerExtra := customtypes.GetHeaderExtra(header)
-	// Make sure the block isn't too far in the future
-	if maxBlockTime := uint64(currentTime.Add(MaxFutureBlockTime).Unix()); header.Time > maxBlockTime {
-		return fmt.Errorf("%w: %d > allowed %d", ErrBlockTooFarInFuture, header.Time, maxBlockTime)
-	}
-
-	// Validate TimeMilliseconds field isn't present prior to Granite
-	if !extraConfig.IsGranite(header.Time) {
-		if headerExtra.TimeMilliseconds != nil {
-			return ErrTimeMillisecondsBeforeGranite
-		}
-		return nil
-	}
-
-	// After Granite, TimeMilliseconds is required
-	if headerExtra.TimeMilliseconds == nil {
-		return ErrTimeMillisecondsRequired
-	}
-
-	// Validate TimeMilliseconds is consistent with header.Time
-	if header.Time != *headerExtra.TimeMilliseconds/1000 {
-		return fmt.Errorf("%w: header.Time (%d) != TimeMilliseconds/1000 = (%d)",
-			ErrTimeMillisecondsMismatched,
-			header.Time,
-			*headerExtra.TimeMilliseconds/1000,
-		)
-	}
-
-	// TODO: If the parent is also after Granite, ensure time is strictly increasing
 	return nil
 }
