+++ plugin/evm/customheader/extra_test.go
@@ -6,7 +6,6 @@ package customheader
 import (
 	"math/big"
 	"testing"
-	"time"
 
 	"github.com/ava-labs/avalanchego/vms/components/gas"
 	"github.com/ava-labs/libevm/core/types"
@@ -608,78 +607,6 @@ func TestVerifyExtra(t *testing.T) {
 		t.Run(test.name, func(t *testing.T) {
 			err := VerifyExtra(test.rules, test.extra)
 			require.ErrorIs(t, err, test.expected)
-		})
-	}
-}
-
-func TestVerifyTime(t *testing.T) {
-	var (
-		time        = time.Unix(1714339200, 123_456_789)
-		timeSeconds = uint64(time.Unix())
-		timeMillis  = uint64(time.UnixMilli())
-	)
-	tests := map[string]struct {
-		timeSeconds      uint64
-		timeMilliseconds *uint64
-		extraConfig      *extras.ChainConfig
-		expectedErr      error
-	}{
-		"pre_granite_time_milliseconds_should_fail": {
-			timeSeconds:      timeSeconds,
-			timeMilliseconds: utils.NewUint64(timeMillis),
-			extraConfig:      extras.TestFortunaChainConfig,
-			expectedErr:      ErrTimeMillisecondsBeforeGranite,
-		},
-		"pre_granite_time_nil_milliseconds_should_work": {
-			timeSeconds:      timeSeconds,
-			timeMilliseconds: nil,
-			extraConfig:      extras.TestFortunaChainConfig,
-			expectedErr:      nil,
-		},
-		"granite_time_milliseconds_should_be_non_nil_and_fail": {
-			timeSeconds:      timeSeconds,
-			timeMilliseconds: nil,
-			extraConfig:      extras.TestGraniteChainConfig,
-			expectedErr:      ErrTimeMillisecondsRequired,
-		},
-		"granite_time_milliseconds_matching_time_should_work": {
-			timeSeconds:      timeSeconds,
-			timeMilliseconds: utils.NewUint64(timeSeconds * 1000),
-			extraConfig:      extras.TestGraniteChainConfig,
-			expectedErr:      nil,
-		},
-		"granite_time_milliseconds_matching_time_rounded_should_work": {
-			timeSeconds:      timeSeconds,
-			timeMilliseconds: utils.NewUint64(timeMillis),
-			extraConfig:      extras.TestGraniteChainConfig,
-			expectedErr:      nil,
-		},
-		"granite_time_milliseconds_less_than_time_should_fail": {
-			timeSeconds:      timeSeconds,
-			timeMilliseconds: utils.NewUint64(timeSeconds*1000 - 1),
-			extraConfig:      extras.TestGraniteChainConfig,
-			expectedErr:      ErrTimeMillisecondsMismatched,
-		},
-		"granite_time_milliseconds_greater_than_time_should_fail": {
-			timeSeconds:      timeSeconds,
-			timeMilliseconds: utils.NewUint64(timeSeconds * 1001),
-			extraConfig:      extras.TestGraniteChainConfig,
-			expectedErr:      ErrTimeMillisecondsMismatched,
-		},
-	}
-
-	for name, test := range tests {
-		t.Run(name, func(t *testing.T) {
-			header := customtypes.WithHeaderExtra(
-				&types.Header{
-					Time: test.timeSeconds,
-				},
-				&customtypes.HeaderExtra{
-					TimeMilliseconds: test.timeMilliseconds,
-				},
-			)
-			err := VerifyTime(test.extraConfig, header, time)
-			require.ErrorIs(t, err, test.expectedErr)
 		})
 	}
 }
